fix: address hosted review wait feedback
